+++ changes.md
@@ -1,5 +1,6 @@
 #### 0.0.2
 * FIX: improved compatibility with Stateless mode.
+* FIX: add support for Elementor Pro forms.
 * FIX: get the correct list of files for Compatibility Files Sync.
 
 #### 0.0.1